la til flagg på mpi.c
har fikset alle feil jeg fikk

+++ oving4/mpi.com.h
@@ -2,13 +2,13 @@
 #include "mpi.com.private.h"
 #include <stdio.h>
 struct mpi_com {
-	int nprocs;
-	int rank;
+	size_t nprocs;
+	size_t rank;
 } uplink;
 
 void mpi_com_Init(int *argc, char ***argv);
 void mpi_com_Finalize(void);
-void generateSendCounts( int * sendCounts , int length);
+void generateSendCounts( int * sendCounts , size_t length);
 void generateSendDisplacements( int *sendDisplacements, int *sendCounts );
 void printCountsAndDisplacements( int * sendCounts, int * sendDisplacements);
 double reducePlus(double *Vector, int length ) ;
